+++ src/pharmpy/modeling/odes.py
@@ -326,7 +326,7 @@ def set_bolus_absorption(model):
     return model
 
 
-def seq_zo_fo_absorption(model):
+def set_seq_zo_fo_absorption(model):
     """Set or change to sequential zero order first order absorption rate. Initial estimate for
     absorption rate is set the previous rate if available, otherwise it is set to the time of
     first observation/2 is used.