+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ next version
 * Rename zero_order_absorption to set_zero_order_absorption
 * Rename first_order_absorption to set_first_order_absorption
 * Rename bolus_absorption to set_bolus_absorption
+* Rename seq_zo_fo_absorption to set_seq_zo_fo_absorption
 
 0.28.0 (2021-08-24)
 -------------------